Added unit field to tables
The tables now have a field for units and the framework exists to put unit into the DB and get it out. Lupascu, you need to add the GUI Labels for users to put a value for unit. Then, adjust your HTML generation to take into account this new piece of data.